update parent_backup_dirs default location
Default the parent_backup_dirs location to the parent directory of the
primariy data directory on each host.

The "parent_backup_dirs" parameter sets the internal location to store
the backup of the master data directory and user defined master
tablespaces. It defaults to the parent directory of each primary data
directory on each host including the standby. For example,
/data/coordinator given /data/coordinator/gpseg-1, and /data1/primaries
given /data1/primaries/gpseg1.

The parent_backup_dirs parameter accepts either a single directory or
multiple host:directory pairs. To specify a single directory across all
hosts set a single directory such as /dir. To specify different
directories for each host use the form
"host1:/dir1,host2:/dir2,host3:/dir3" where the first host must be the
master.

+++ hub/delete_backup_directories.go
@@ -11,16 +11,18 @@ import (
 	"github.com/greenplum-db/gpupgrade/upgrade"
 )
 
-func DeleteBackupDirectories(streams step.OutStreams, agentConns []*idl.Connection, backupDir string) error {
-	// delete on coordinator
-	err := upgrade.DeleteDirectories([]string{backupDir}, []string{}, streams)
+func DeleteBackupDirectories(streams step.OutStreams, agentConns []*idl.Connection, backupDirs BackupDirs) error {
+	err := upgrade.DeleteDirectories([]string{backupDirs.CoordinatorBackupDir}, []string{}, streams)
 	if err != nil {
 		return err
 	}
 
-	// delete on segments
 	request := func(conn *idl.Connection) error {
-		req := &idl.DeleteBackupDirectoryRequest{BackupDir: backupDir}
+		if _, ok := backupDirs.AgentHostsToBackupDir[conn.Hostname]; !ok {
+			return nil
+		}
+
+		req := &idl.DeleteBackupDirectoryRequest{BackupDir: backupDirs.AgentHostsToBackupDir[conn.Hostname]}
 		_, err := conn.AgentClient.DeleteBackupDirectory(context.Background(), req)
 		return err
 	}
